+++ consumer/src/main.ts
@@ -3,6 +3,7 @@ import { aim } from "./aim"
 import { attributes } from "./attributes"
 import { elements } from "./elements"
 import { intro } from "./intro"
+import { mixedContent } from "./mixed-content"
 import { nested } from "./nested"
 import { reactive } from "./reactive"
 
@@ -13,6 +14,7 @@ div(
     intro(),
     aim(),
     nested(),
+    mixedContent(),
     attributes(),
     elements(),
     reactive(),